+++ .vscode/settings.json
@@ -5,7 +5,12 @@
 	"python.testing.unittestEnabled": false,
 	"python.testing.pytestEnabled": true,
 	"python.analysis.typeCheckingMode": "basic",
-	"python.formatting.provider": "black",
+	"python.formatting.provider": "none",
 	"python.languageServer": "Pylance",
-	"rust-analyzer.linkedProjects": ["./manager/Cargo.toml"]
+	"rust-analyzer.linkedProjects": [
+		"./manager/Cargo.toml"
+	],
+	"[python]": {
+		"editor.defaultFormatter": "ms-python.black-formatter"
+	}
 }
