+++ core/optimizations/trace_utils.py
@@ -1,8 +1,15 @@
 from typing import Tuple
+import logging
+import warnings
 
 from diffusers.models.unet_2d_condition import UNet2DConditionOutput
 
 import torch
+from tqdm.auto import tqdm
+
+from core.config import config
+
+logger = logging.getLogger(__name__)
 
 
 def generate_inputs(
@@ -33,3 +40,45 @@ class TracedUNet(torch.nn.Module):
 
         sample = self.og.forward(latent_model_input, t, encoder_hidden_states)[0]
         return UNet2DConditionOutput(sample=sample)
+
+
+def warmup(
+    model: torch.nn.Module, amount: int, dtype: torch.dtype, device: torch.device
+) -> None:
+    "Warms up model with amount generated sample inputs."
+
+    model.eval()
+    with torch.inference_mode():
+        for _ in tqdm(range(amount), unit="it", desc="Warming up", unit_scale=False):
+            model(*generate_inputs(dtype, device))
+
+
+def trace_model(
+    model: torch.nn.Module,
+    dtype: torch.dtype,
+    device: torch.device,
+    iterations: int = 25,
+    ipex: bool = False,
+) -> torch.nn.Module:
+    "Traces the model for inference"
+
+    og = model
+    from functools import partial
+
+    model.forward = partial(model.forward, return_dict=False)
+    warmup(model, iterations, dtype, device)
+    if config.api.channels_last and not ipex:
+        model.to(memory_format=torch.channels_last)  # type: ignore
+    logger.debug("Starting trace")
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore")
+        if config.api.device_type == "cpu":
+            torch.jit.enable_onednn_fusion(True)
+        model = torch.jit.trace(model, generate_inputs(dtype, device), check_trace=False)  # type: ignore
+        model = torch.jit.freeze(model)  # type: ignore
+    logger.debug("Tracing done")
+    warmup(model, iterations // 5, dtype, device)
+
+    rn = TracedUNet(model)
+    del og
+    return rn
